+++ src/Pororoca.Desktop/Behaviors/ResponseCapturesDataGridDropHandler.cs
@@ -1,0 +1,84 @@
+using Avalonia.Controls;
+using Avalonia.Input;
+using Avalonia.VisualTree;
+using Avalonia.Xaml.Interactions.DragAndDrop;
+using Pororoca.Desktop.ViewModels.DataGrids;
+
+namespace Pororoca.Desktop.Behaviors;
+
+public interface IResponseCapturesDataGridOwner
+{
+    HttpResponseCapturesDataGridViewModel ResCapturesTableVm { get; }
+}
+
+public class ResponseCapturesDataGridDropHandler : DropHandlerBase
+{
+    private bool Validate(DataGrid dg, DragEventArgs e, object? sourceContext, object? targetContext, bool bExecute)
+    {
+        if (sourceContext is not HttpResponseCaptureViewModel sourceItem
+         || targetContext is not IResponseCapturesDataGridOwner vm
+         || dg.GetVisualAt(e.GetPosition(dg)) is not Control targetControl
+         || targetControl.DataContext is not HttpResponseCaptureViewModel targetItem)
+        {
+            return false;
+        }
+
+        var items = vm.ResCapturesTableVm.Items;
+        int sourceIndex = items.IndexOf(sourceItem);
+        int targetIndex = items.IndexOf(targetItem);
+
+        if (sourceIndex < 0 || targetIndex < 0)
+        {
+            return false;
+        }
+
+        switch (e.DragEffects)
+        {
+            case DragDropEffects.Copy:
+            {
+                if (bExecute)
+                {
+                    HttpResponseCaptureViewModel clone = new(items, sourceItem.ToResponseCapture().Copy());
+                    InsertItem(items, clone, targetIndex + 1);
+                }
+                return true;
+            }
+            case DragDropEffects.Move:
+            {
+                if (bExecute)
+                {
+                    MoveItem(items, sourceIndex, targetIndex);
+                }
+                return true;
+            }
+            case DragDropEffects.Link:
+            {
+                if (bExecute)
+                {
+                    SwapItem(items, sourceIndex, targetIndex);
+                }
+                return true;
+            }
+            default:
+                return false;
+        }
+    }
+        
+    public override bool Validate(object? sender, DragEventArgs e, object? sourceContext, object? targetContext, object? state)
+    {
+        if (e.Source is Control && sender is DataGrid dg)
+        {
+            return Validate(dg, e, sourceContext, targetContext, false);
+        }
+        return false;
+    }
+
+    public override bool Execute(object? sender, DragEventArgs e, object? sourceContext, object? targetContext, object? state)
+    {
+        if (e.Source is Control && sender is DataGrid dg)
+        {
+            return Validate(dg, e, sourceContext, targetContext, true);
+        }
+        return false;
+    }
+}