+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Show WebSocket connection response HTTP status code and headers by [@tetropolix](https://github.com/tetropolix).
 * Import OpenAPI / Swagger files.
 * Re-run response captures (no need to send a request again).
+* Drag-and-drop in collection and environment variables, and response captures (for other tables, will be added in the next releases).
 
 ### Bug Fixes